// app/interfaces/customSession.ts
import { DefaultSession } from "next-auth";

export interface CustomSession extends DefaultSession {
    accessToken?: string; // Add the access token
    user?: {
        id: string;
        firstName: string;
        lastName: string;
        subject: string;
        role?: string; // Optional roles array
        isActive?: string; // Optional permissions array
        isBlocked?: string;
        image?: string
    };
}